cambiando de integer e increments a biginteger y bigIncrements

+++ src/database/migrations/2017_09_07_043330_migration_roles.php
@@ -15,36 +15,36 @@ class MigrationRoles extends Migration
 	public function up() 
 	{
 		Schema::create('roles', function (Blueprint $table){
-			$table->increments('rolid');
+			$table->bigIncrements('rolid');
 			$table->string('nombre');
 			$table->string('descripcion');
 			$table->nullableTimestamps();
 		});
 		Schema::create('modulos', function (Blueprint $table){
-			$table->increments('moduloid');
+			$table->bigIncrements('moduloid');
 			$table->string('nombre');
 			$table->string('ruta');
 			$table->nullableTimestamps();
 		});
 		Schema::create('permisos', function (Blueprint $table){
-			$table->increments('permisoid');
+			$table->bigIncrements('permisoid');
 			$table->string('nombre');
 			$table->string('nombreLaravel');
 			$table->string('color');
 			$table->nullableTimestamps();
 		});
 		Schema::create('moduloPermiso', function (Blueprint $table){
-			$table->increments('modulopermisoid');
-			$table->integer('moduloid')->unsigned();
-			$table->integer('permisoid')->unsigned();
+			$table->bigIncrements('modulopermisoid');
+			$table->bigInteger('moduloid')->unsigned();
+			$table->bigInteger('permisoid')->unsigned();
 			$table->foreign('permisoid')->references('permisoid')->on('permisos');
 			$table->foreign('moduloid')->references('moduloid')->on('modulos');
 			$table->nullableTimestamps();
 		});
 		Schema::create('rolModuloPermiso', function (Blueprint $table){
-			$table->increments('rolmodulopermisoid');
-			$table->integer('rolid')->unsigned();
-			$table->integer('modulopermisoid')->unsigned();
+			$table->bigIncrements('rolmodulopermisoid');
+			$table->bigInteger('rolid')->unsigned();
+			$table->bigInteger('modulopermisoid')->unsigned();
 			$table->foreign('rolid')->references('rolid')->on('roles');
 			$table->foreign('modulopermisoid')->references('modulopermisoid')->on('moduloPermiso');
 			$table->nullableTimestamps();
