+++ src/database/migrations/2017_09_09_075130_migrationUsuarioRol.php
@@ -14,9 +14,9 @@ class MigrationUsuarioRol extends Migration
 	public function up() 
 	{
 		Schema::create('usuarioRol', function (Blueprint $table){
-			$table->increments('usuariorolid');
-			$table->integer('rolid')->unsigned();
-			$table->integer('usuarioid')->unsigned();
+			$table->bigIncrements('usuariorolid');
+			$table->bigInteger('rolid')->unsigned();
+			$table->bigInteger('usuarioid')->unsigned();
 			$table->foreign('rolid')->references('rolid')->on('roles');
 			$table->nullableTimestamps();
 		});
